requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# LLM providers via LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# Embeddings & vector search
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0

# Fuzzy matching
rapidfuzz>=3.0.0

# Config & utilities
pyyaml>=6.0
pydantic>=2.0.0
numpy>=1.24.0
